% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_spanish.R
\name{acc_rm}
\alias{acc_rm}
\title{Remove accents.}
\usage{
acc_rm(x, id)
}
\arguments{
\item{x}{character.}

\item{id}{a single string with transform identifier, see stringi::stri_trans_list.}
}
\description{
Remove accents.
}
