% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_util_function.R
\name{add_util_function}
\alias{add_util_function}
\title{Title add_util_function}
\usage{
add_util_function(.fn, .pkg = "~/Documents/pkg_dev/myutilities/", print = TRUE)
}
\description{
Title add_util_function
}
