% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_cat.R
\name{age_cat}
\alias{age_cat}
\title{function to categorize ages
From http://www.r-bloggers.com/r-function-of-the-day-cut}
\usage{
age_cat(x, lower = 0, upper, by = 10, sep = "-", above.char = "+", style = 1)
}
\arguments{
\item{x:}{Vector with age as numeric.}

\item{lower:}{int lower age}

\item{upper:}{int upper age}

\item{by:}{int every}

\item{sep:}{char default "-".}

\item{above.char:}{default "+".}

\item{style:}{1 = '50 - 54', 2 = '50', 3 = '11'}
}
\description{
function to categorize ages
From http://www.r-bloggers.com/r-function-of-the-day-cut
}
