% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_fdates.R
\name{age_fdates}
\alias{age_fdates}
\title{Calculate age from dates.}
\usage{
age_fdates(from, to = as.Date(lubridate::today()), units = "years",
  floor = FALSE)
}
\arguments{
\item{from}{first date. Date object}

\item{to}{second date. Default as.Date(today()).}

\item{units}{years, months, days or hours. Default years.}

\item{floor}{... Default FALSE.}
}
\description{
Calculate age from dates.
}
