% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_helpers.R
\name{assignment_helpers}
\alias{assignment_helpers}
\alias{m_assign}
\alias{\%<<-\%}
\title{Assignment helpers}
\usage{
m_assign(x, value, env = .GlobalEnv)

 \%<<-\% 
}
\arguments{
\item{x}{vactor of symbols for assignment. Same as lhs of \verb{\%<<-\%}.}

\item{value}{values to assign.}

\item{env}{environment.}
}
\description{
Assignment helpers
}
