% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{cb_copy}
\alias{cb_copy}
\title{Copy data.frames, tables, xtabs, ftables, tbl_df to clipboard prepared for MS Excel.}
\usage{
cb_copy(x, ...)
}
\arguments{
\item{x:}{data.frame, table, xtabs, ftable or tbl_df.}
}
\description{
Copy data.frames, tables, xtabs, ftables, tbl_df to clipboard prepared for MS Excel.
}
