% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{cb_copyunix}
\alias{cb_copyunix}
\title{Copy an object in the clipboard}
\usage{
cb_copyunix(x, sep = "\\t", dec = ",", row.names = FALSE,
  col.names = TRUE)
}
\arguments{
\item{sep}{The object to copy.}

\item{row.names}{Copy row names (default is FALSE)}

\item{col.names}{Copy column names (default is TRUE)}

\item{sep}{A character to be used as separator for each column of the object}
}
\value{
copy the object as character in the clipboard
}
\description{
Copy an object in the clipboard
}
\author{
freecube source:http://stackoverflow.com/questions/10959521/how-to-write-to-clipboard-on-ubuntu-linux-in-r
}
