% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unquoted.R
\name{cc}
\alias{cc}
\title{Concatenate unquoted object names}
\usage{
cc(...)
}
\arguments{
\item{...}{quoted or unquoted object names}
}
\value{
a character vector.
}
\description{
Concatenate unquoted object names
}
