% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classint.R
\name{classint}
\alias{classint}
\title{Create vector with Interval Class from classIntervals functions}
\usage{
classint(var, n, style = "quantile", labels = NULL,
  include.lowest = T)
}
\arguments{
\item{var}{numeric: variable to apply class intervals.}

\item{n}{int: number of intervals.}

\item{style}{chr: breaks style. chosen style: one of "fixed", "sd", "equal", "pretty", "quantile", "kmeans", "hclust", "bclust", "fisher", or "jenks". Default: "quantile".}

\item{labels:}{chr vector with labels for class.}

\item{include.lowest:}{logical, indicating if an equal to the lowest}
}
\value{
factor.
}
\description{
Create vector with Interval Class from classIntervals functions
}
