% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_c.R
\name{compare_c}
\alias{compare_c}
\title{Function for compare vectors with a easy view. Shows elements present, absent and outer.}
\usage{
compare_c(c_to_compare, c_standard)
}
\arguments{
\item{c_to_compare}{Vector to compare elements.}

\item{c_standard}{Vector with standard elements.}
}
\description{
Function for compare vectors with a easy view. Shows elements present, absent and outer.
}
