% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consec_date.R
\name{consec_date}
\alias{consec_date}
\title{This function iterates through values using a sliding window, 
applying a condition and evaluate it to each sub-window 
then return the first date of the windows that match condition}
\usage{
consec_date(date, value, x, .after = 7)
}
\arguments{
\item{date}{date. i-ndex date.}

\item{value}{numeric. Count values.}

\item{x}{numeric. condition that is more than x value.}

\item{.after}{The number of values after the current element of .i to include 
in the sliding window. Set to Inf to select all elements after the 
current element. Negative values are allowed, which allows you 
to "look backwards".}
}
\value{
date.
}
\description{
This function iterates through values using a sliding window, 
applying a condition and evaluate it to each sub-window 
then return the first date of the windows that match condition
}
