% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consec_idx.R
\name{consec_idx}
\alias{consec_idx}
\title{This function iterates through values using a sliding window, 
applying a condition and evaluate it to each sub-window 
then return the first or last index of the windows that match condition}
\usage{
consec_idx(index, value, expr, .after = 7, .f = "min", .messages = F)
}
\arguments{
\item{index}{date. i-ndex date.}

\item{value}{numeric. Count values.}

\item{expr}{expression. type of .x > 2.}

\item{.after}{The number of values after the current element of .i to include 
in the sliding window. Set to Inf to select all elements after the 
current element. Negative values are allowed, which allows you 
to "look backwards".}

\item{.f}{function to apply to find min o max of all windows.}

\item{messages}{show messages for more information about the windows.}
}
\value{
window's vector.
}
\description{
This function iterates through values using a sliding window, 
applying a condition and evaluate it to each sub-window 
then return the first or last index of the windows that match condition
}
