% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_labs.R
\name{create_labs}
\alias{create_labs}
\title{Create labels from numeric intervals.}
\usage{
create_labs(lower, upper, by, above.char, sep)
}
\arguments{
\item{lower:}{int lower.}

\item{upper:}{int upper.}

\item{by:}{int every.}

\item{sep:}{char default "-".}

\item{above.char:}{char default "+".}
}
\value{
character vector.
}
\description{
Create labels from numeric intervals.
}
