% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInfo.R
\name{dataInfo}
\alias{dataInfo}
\title{Get detailed information of a data.frame}
\usage{
dataInfo(dataFrame)
}
\arguments{
\item{dataFrame}{A data.frame object.}
}
\value{
A data.frame with summary information of the input `dataFrame`
}
\description{
Get detailed information of a data.frame
}
