% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateToInt.R
\name{DateToInt}
\alias{DateToInt}
\title{Calculate interval unit from dates.}
\usage{
DateToInt(from, to = as.Date(lubridate::today()), units = "years",
  floor = FALSE)
}
\arguments{
\item{from}{first date. Date object}

\item{to}{second date. Default as.Date(today()).}

\item{units}{years, months, days or hours. Default years.}

\item{floor}{... Default FALSE.}
}
\description{
Calculate interval unit from dates.
}
