% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drecode.R
\name{drecode}
\alias{drecode}
\title{Recode throw traduction table. Check for inconsistency.}
\usage{
drecode(x, y, varx, vary, field, name = lab)
}
\arguments{
\item{x}{data.frame with the field to recode.}

\item{y}{data.frame traduction table.}

\item{varx}{variable to recode.}

\item{vary:}{variable to link.}

\item{field:}{variable/s recoded.}

\item{name:}{name/s of field recoded. Default lab.}
}
\value{
data.frame with \code{x} fields and recoded field with \code{name}.
}
\description{
Recode throw traduction table. Check for inconsistency.
}
