% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidates.R
\name{dx_toNumber}
\alias{dx_toNumber}
\title{Return excel number from date.}
\usage{
dx_toNumber(x)
}
\arguments{
\item{date:}{Date vector}
}
\value{
numeric: number
}
\description{
Return excel number from date.
}
