% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidates.R
\name{epidates_date2week}
\alias{epidates_date2week}
\title{Return epi - week from date. Weeks start on sunday.}
\usage{
epidates_date2week(date)
}
\arguments{
\item{date:}{Date object}
}
\value{
numeric: week number
}
\description{
Return epi - week from date. Weeks start on sunday.
}
