% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidates.R
\name{epidates_yw2month}
\alias{epidates_yw2month}
\title{Return month from year and week}
\usage{
epidates_yw2month(year, week)
}
\arguments{
\item{year:}{int year with format YYYY.}

\item{week:}{int week number.}
}
\value{
month as integer.
}
\description{
Return month from year and week
}
