% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiwk_funs.R
\name{epiwk_asDate}
\alias{epiwk_asDate}
\title{Funciones para trabajar con semanas epidemiológicas
This function returns the first day of the week in date data type 
when the year and week are entered in integer format.}
\usage{
epiwk_asDate(year, week)
}
\arguments{
\item{year}{<int> year}

\item{week}{<int> week}
}
\value{
<date> first day of the week
}
\description{
Funciones para trabajar con semanas epidemiológicas
This function returns the first day of the week in date data type 
when the year and week are entered in integer format.
}
