% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiwk_funs.R
\name{epiwk_calendar}
\alias{epiwk_calendar}
\title{This function returns a calendar}
\usage{
epiwk_calendar(year)
}
\arguments{
\item{year}{<int> year}
}
\value{
<tibble> tibble
}
\description{
This function returns a calendar
}
