% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiwk_funs.R
\name{epiwk_floor}
\alias{epiwk_floor}
\title{Return the first day of the week that x <date> is in.}
\usage{
epiwk_floor(x)
}
\arguments{
\item{x}{<date>}
}
\value{
First day of the week
}
\description{
Return the first day of the week that x <date> is in.
}
