% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiwk_funs.R
\name{epiwk_int}
\alias{epiwk_int}
\title{Function epiwk_int}
\usage{
epiwk_int(year, weeks)
}
\arguments{
\item{year}{<int> year}

\item{week}{<int> week}
}
\value{
<date> vector day of the week
}
\description{
This function returns a vector with the dates of the week in date data type 
when the year and week are entered in integer format.
}
