% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_exp.R
\name{filter_exp}
\alias{filter_exp}
\title{Filter expressions. Helpful for pre estableshing expressions from the package.
cie10 filter:
tb = tuberculosis.
fq = Cystic Fibrosis.}
\usage{
filter_exp(x, cond, ...)
}
\arguments{
\item{x:}{data.frame}

\item{cond:}{expression. Ex: ~codmuer%ilike%"M%%". See my_filter_list for preestablished filters.}

\item{...:}{variables names. Ex: codmuer = cie10.}
}
\description{
Filter expressions. Helpful for pre estableshing expressions from the package.
cie10 filter:
tb = tuberculosis.
fq = Cystic Fibrosis.
}

