% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_spanish.R
\name{format_sp}
\alias{format_sp}
\title{Spanish format print for labels and plus flag.}
\usage{
format_sp(x, digits = 2, dec = ",", big.mark = ".", plus.flag = F,
  flag = NULL)
}
\arguments{
\item{x}{numeric to format spanish type}

\item{digits}{number of digits. Default 3.}

\item{dec}{decimal mark. Default comma.}

\item{big.mark}{mark every 3 numbers.}

\item{plus.flag}{logical. If \code{TRUE} add plus sign.}

\item{flag}{add a right flag. Default NULL. Useful add \code{\%}.}
}
\value{
character with spanish format and rounded by digits.
}
\description{
Spanish format print for labels and plus flag.
}
