% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_then.R
\name{if_then}
\alias{if_then}
\alias{\%||\%}
\alias{if_null}
\alias{\%0\%}
\alias{if_empty}
\alias{\%na\%}
\alias{if_na}
\title{If x is null/empty/NA then y otherwise x}
\usage{
x \%||\% y

if_null(x, y)

x \%0\% y

if_empty(x, y)

x \%na\% y

if_na(x, y)
}
\arguments{
\item{x}{original.}

\item{y}{alternative}
}
\value{
x or y.
}
\description{
If x is null/empty/NA then y otherwise x

NULL ----

empty ----

NA ----
}
