% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myutilities-package.R
\docType{package}
\name{myutilities-package}
\alias{myutilities}
\alias{myutilities-package}
\title{myutilities: Collection of functions that I use frequently.}
\description{
Functions that have been useful to me in my day-to-day work in epidemiology
}
\keyword{internal}
