% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_list.R
\name{named_list}
\alias{named_list}
\title{Title named_list}
\usage{
named_list(
  ...,
  .named = TRUE,
  .ignore_empty = c("trailing", "none", "all"),
  .preserve_empty = FALSE,
  .homonyms = c("keep", "first", "last", "error"),
  .check_assign = FALSE
)
}
\description{
Title named_list
}
