% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outersect.R
\name{outersect}
\alias{outersect}
\title{outersect: The opposite of R's intersect function.
Author: Tony Breyal example from http://www.r-bloggers.com/outersect-the-opposite-of-rs-intersect-function/.}
\usage{
outersect(x, y, ...)
}
\arguments{
\item{x:}{vector to compare.}

\item{y:}{vector to compare.}

\item{...:}{vectors to compare.}
}
\value{
vector: Set of elements which are in either of the sets and not in their intersection.
}
\description{
outersect: The opposite of R's intersect function.
Author: Tony Breyal example from http://www.r-bloggers.com/outersect-the-opposite-of-rs-intersect-function/.
}
