% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{pathPrep}
\alias{pathPrep}
\title{function that read path on windows from clipboard prepares it for R and paste it in clipboard.
Copy path
directory <- pathPrep() or paste it on script}
\usage{
pathPrep(path = "clipboard")
}
\description{
function that read path on windows from clipboard prepares it for R and paste it in clipboard.
Copy path
directory <- pathPrep() or paste it on script
}

