% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_se.R
\name{pct_se}
\alias{pct_se}
\title{Standard Error from percentege.}
\usage{
pct_se(pct, n, alpha = 0.05)
}
\arguments{
\item{pct:}{percentage.}

\item{n:}{length.}

\item{alpha:}{alpha error.}
}
\description{
Standard Error from percentege.
}
