% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodFromDates.R
\name{periodFromDates}
\alias{periodFromDates}
\title{Period between two dates with format}
\usage{
periodFromDates(from, to, format = "\%2d years and \%2d months")
}
\arguments{
\item{from}{<date>}

\item{to}{<date>}
}
\value{
chr variable with format
}
\description{
Period between two dates with format
}
