% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_export.R
\name{pptx_export}
\alias{pptx_export}
\title{Export plot to pptx}
\usage{
pptx_export(plot, template = file.choose(), design_name,
  file = "test plot.pptx")
}
\arguments{
\item{plot:}{plot to export}

\item{template:}{file pptx with the design.}

\item{design_name:}{name of design.}

\item{file:}{export file name.}
}
\description{
Export plot to pptx
}
