% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers from a vector}
\usage{
remove_outliers(x, na.rm = TRUE, ...)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
This function takes a vector and removes outliers out of 1.5 IQR
}
