% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_acc.R
\name{rm_acc}
\alias{rm_acc}
\title{Remove accents.}
\usage{
rm_acc(x, id)
}
\arguments{
\item{x}{character.}

\item{id}{a single string with transform identifier, see stringi::stri_trans_list. Default "Latin-ASCII"}
}
\description{
Remove accents.
}
