% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapply_df.R
\name{sapply_df}
\alias{sapply_df}
\title{Applying multiple functions to data.frame.
Extracted from https://www.r-bloggers.com/applying-multiple-functions-to-data-frame/}
\usage{
sapply_df(...)
}
\arguments{
\item{...:}{Data frame, functions names or unnamed functions.}
}
\description{
Applying multiple functions to data.frame.
Extracted from https://www.r-bloggers.com/applying-multiple-functions-to-data-frame/
}
