% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_brewer_pal.R
\name{scale_brewer_pal}
\alias{scale_brewer_pal}
\title{Create palette Color Brewer from factor levels.}
\usage{
scale_brewer_pal(x, palette = "Blues")
}
\arguments{
\item{x}{factor or chr vector}

\item{palette}{chr: "Blues", "Reds", "Greens","Greys". Default: "Blues".}
}
\value{
chr vector.
}
\description{
Create palette Color Brewer from factor levels.
}
