% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str2vec_class.R
\name{str2vec_class}
\alias{str2vec_class}
\title{Empty Vectors from character class definition}
\usage{
str2vec_class(x = character())
}
\arguments{
\item{x}{string with class name.}
}
\value{
empty vector with class \code{x}.
}
\description{
Empty Vectors from character class definition
}
