% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_between.R
\name{str_between}
\alias{str_between}
\title{Do values in a string vector fall in specified range?
str_between is case insensitive by default.}
\usage{
str_between(x, left, right, exclude = NULL)
}
\arguments{
\item{x}{A string vector of values.}

\item{left, right}{Boundary values.}

\item{exclude}{vector string exclude values inside range.}
}
\description{
Do values in a string vector fall in specified range?
str_between is case insensitive by default.
}
