% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_between_family.R
\name{str_between_factory}
\alias{str_between_factory}
\title{Function str between family
Function factory 
Make the str_between family}
\usage{
str_between_factory(fns)
}
\arguments{
\item{chr}{function name}
}
\description{
Function str between family
Function factory 
Make the str_between family
}
