% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_all.R
\name{unnest_all}
\alias{unnest_all}
\title{Unnesting all columns}
\usage{
unnest_all(data, ...)
}
\arguments{
\item{data}{data.frame or object coercible to it.}

\item{...}{unused for now.}
}
\value{
data.frame
}
\description{
Unnesting all columns
}
