% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vap.R
\name{vap.data.frame}
\alias{vap.data.frame}
\title{Annual percent change
For use with do function from dplyr.
df %>% group_by(group) %>% do(vap(., year, rate))
Some problems with Inf values in y were observed.}
\usage{
\method{vap}{data.frame}(df, x, y, tests = F)
}
\arguments{
\item{tests}{logical: if TRUE computes p-value for lineal model tests. See Details.}

\item{df:}{data.frame with data.}

\item{x:}{year}

\item{y:}{rate}
}
\value{
data.frame. vap: annual percent change. lwr: lower IC 95\code{\%}. upr: upper IC 95\code{\%}.
}
\description{
Annual percent change
For use with do function from dplyr.
df %>% group_by(group) %>% do(vap(., year, rate))
Some problems with Inf values in y were observed.
}
\details{
Lineal model tests: SW = residual normality by Shapiro Wilks. BP: Performs the Breusch-Pagan test against heteroskedasticity. HC: Harvey-Collier test for linearity. DW: Computes bootstrapped p-values for residual autocorrelations and generalized Durbin-Watson statistics.
}
