% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vap.R
\name{vap.numeric}
\alias{vap.numeric}
\title{Annual percent change}
\usage{
\method{vap}{numeric}(x, y, tests = FALSE)
}
\arguments{
\item{tests}{logical: if TRUE computes p-value for lineal model tests. See Details.}

\item{x:}{vector}

\item{y:}{vector}
}
\value{
vector. vap: annual percent change. lwr: lower IC 95%. upr: upper IC 95%.
}
\description{
Annual percent change
}
\details{
Lineal model tests: SW = residual normality by Shapiro Wilks. BP: Performs the Breusch-Pagan test against heteroskedasticity. HC: Harvey-Collier test for linearity. DW: Computes bootstrapped p-values for residual autocorrelations and generalized Durbin-Watson statistics.
}

