% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write2txt.R
\name{write.delim2}
\alias{write.delim2}
\title{Function to write in spanish format txt files. Consistent with read.delim2. Useful for read in excel with spansh configuration.}
\usage{
write.delim2(x, dir = choose.files())
}
\arguments{
\item{x}{data.frame to write.}

\item{dir}{path directory and name of file with extension. Default open windows to choose file localization.}

\item{ext}{extension file. Default \code{txt}.}
}
\value{
plain text file with header, sep by tabular and comma as decimal. NA are empty chr.
}
\description{
Function to write in spanish format txt files. Consistent with read.delim2. Useful for read in excel with spansh configuration.
}
