% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlist.R
\name{xlist}
\alias{xlist}
\title{Title xlist}
\usage{
xlist(
  ...,
  .named = TRUE,
  .ignore_empty = c("trailing", "none", "all"),
  .preserve_empty = FALSE,
  .homonyms = c("keep", "first", "last", "error"),
  .check_assign = FALSE,
  preserve_empty = TRUE
)
}
\description{
Title xlist
}
