% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_else.R
\name{if_then}
\alias{if_then}
\alias{\%||\%}
\alias{\%|0|\%}
\alias{\%na\%}
\title{If x is null/empty/NA then y otherwise x}
\usage{
x \%||\% y

x \%|0|\% y

x \%na\% y
}
\arguments{
\item{x}{original.}

\item{y}{alternative}
}
\value{
x or y.
}
\description{
If x is null/empty/NA then y otherwise x
}
