% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_helpers.R
\name{string_helpers}
\alias{string_helpers}
\alias{no_case}
\title{String helpers}
\usage{
no_case(
  pattern,
  ignore_case = TRUE,
  multiline = FALSE,
  comments = FALSE,
  dotall = FALSE,
  ...
)
}
\arguments{
\item{pattern}{pattern.}

\item{ignore_case}{.}

\item{multiline}{.}

\item{comments}{.}

\item{dotall}{.}

\item{...}{.}
}
\value{
.
}
\description{
String helpers
}
